+++ src/Mapper/Value.php
@@ -14,6 +14,13 @@ final class Value extends StatelessMapper
         return $value;
     }
     
+    protected function buildKeyMapper(iterable $stream): iterable
+    {
+        foreach ($stream as $value) {
+            yield $value => $value;
+        }
+    }
+    
     public function mergeWith(Mapper $other): bool
     {
         return $other instanceof self;